requirements.txt
Flask==0.12.2
gunicorn==19.7.1
tensorflow
numpy
pillow
h5py
